API works receiving an image in base64 encoding - yay

+++ API/classes/ModelInput.cs
@@ -3,5 +3,5 @@ namespace API.classes;
 public class ModelInput : ImageData
 {
     public byte[] Image { get; set; }
-    public UInt16 LabelAsKey { get; set; }
+    public UInt32 LabelAsKey { get; set; }
 }
